8. Adding user Authentication

+++ controllers/expense.js
@@ -4,17 +4,21 @@ exports.addExpense = (req, res) => {
   const amount = req.body.amount;
   const description = req.body.description;
   const category = req.body.category;
+  const id = req.user.id;
+
   Expense.create({
     amount: amount,
     description: description,
     category: category,
+    userId : id,
   })
     .then((result) => res.json(result))
     .catch((err) => console.log(err));
 };
 
 exports.getExpense = (req, res) => {
-  Expense.findAll()
+  const userId = req.user.id;
+  Expense.findAll({ where: { userId: userId } })
     .then((expenses) => {
       res.json(expenses);
     })
@@ -23,7 +27,8 @@ exports.getExpense = (req, res) => {
 
 exports.deleteExpense = (req, res) => {
   const id = req.params.id;
-  Expense.destroy({ where: { id: id } })
+  const userId = req.user.id;
+  Expense.destroy({ where: { id: id, userId: userId } })
     .then((result) => res.json(result))
     .catch((err) => console.log(err));
 };
